requirements.txt
streamlit
pandas
numpy
plotly
requests
groq
python-dotenv
google-api-python-client
matplotlib
streamlit-autorefresh
